+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 * fixing ses secretvalue permissions issue
 * allowing dbadmin username to be customized in DbSecret construct
 * switch from launch configs to launch templates in Asg construct
+* allowing default database to be customized in AuroraCluster construct
 
 # 3.7.0